// src/store/ducks/expenses/expenses.types.ts
/* eslint-disable no-shadow */

export enum ExpenseTypes {
    EXPENSE_CREATE_SUCCESS = 'EXPENSE_CREATE_SUCCESS',
    EXPENSE_CREATE_REQUEST = 'EXPENSE_CREATE_REQUEST',
    EXPENSE_CREATE_FAILURE = 'EXPENSE_CREATE_REQUEST',
    EXPENSE_EDIT_SUCCESS = 'EXPENSE_EDIT_SUCCESS',
    EXPENSE_EDIT_REQUEST = 'EXPENSE_EDIT_REQUEST',
    EXPENSE_EDIT_FAILURE = 'EXPENSE_EDIT_REQUEST',
    EXPENSE_REMOVE_SUCCESS = 'EXPENSE_REMOVE_SUCCESS',
    EXPENSE_REMOVE_REQUEST = 'EXPENSE_REMOVE_REQUEST',
    EXPENSE_REMOVE_FAILURE = 'EXPENSE_REMOVE_REQUEST'
}