// src/store/ducks/payments/payments.types.ts
/* eslint-disable no-shadow */

export enum PaymentTypes {
    PAYMENT_CREATE_SUCCESS = 'PAYMENT_CREATE_SUCCESS',
    PAYMENT_CREATE_REQUEST = 'PAYMENT_CREATE_REQUEST',
    PAYMENT_CREATE_FAILURE = 'PAYMENT_CREATE_REQUEST',
    PAYMENT_EDIT_SUCCESS = 'PAYMENT_EDIT_SUCCESS',
    PAYMENT_EDIT_REQUEST = 'PAYMENT_EDIT_REQUEST',
    PAYMENT_EDIT_FAILURE = 'PAYMENT_EDIT_REQUEST',
    PAYMENT_REMOVE_SUCCESS = 'PAYMENT_REMOVE_SUCCESS',
    PAYMENT_REMOVE_REQUEST = 'PAYMENT_REMOVE_REQUEST',
    PAYMENT_REMOVE_FAILURE = 'PAYMENT_REMOVE_REQUEST'
}