// src/store/ducks/people/people.types.ts
/* eslint-disable no-shadow */

export enum PersonTypes {
    PERSON_CREATE_SUCCESS = 'PERSON_CREATE_SUCCESS',
    PERSON_CREATE_REQUEST = 'PERSON_CREATE_REQUEST',
    PERSON_CREATE_FAILURE = 'PERSON_CREATE_FAILURE',
    PERSON_EDIT_SUCCESS = 'PERSON_EDIT_SUCCESS',
    PERSON_EDIT_REQUEST = 'PERSON_EDIT_REQUEST',
    PERSON_EDIT_FAILURE = 'PERSON_EDIT_FAILURE',
    PERSON_REMOVE_SUCCESS = 'PERSON_REMOVE_SUCCESS',
    PERSON_REMOVE_REQUEST = 'PERSON_REMOVE_REQUEST',
    PERSON_REMOVE_FAILURE = 'PERSON_REMOVE_FAILURE'
}